Docking v2 renaming (#827)
* skip cdd tests without token

* fix running test suite without CDD token

* rename docking stuff

* slightly refactor tests to avoid repetition

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------
