+++ asapdiscovery-data/asapdiscovery/data/tests/test_cdd_download.py
@@ -1,6 +1,7 @@
 """
 Tests for downloading and processing Moonshot data from CDD.
 """
+
 import os
 
 import pytest
@@ -34,6 +35,9 @@ PARSE_ADDED_COLS = [
     "exp_binding_affinity_kcal_mol_95ci_lower",
     "exp_binding_affinity_kcal_mol_95ci_upper",
 ]
+
+# skip tests if no CDD token
+pytestmark = pytest.mark.skipif(not os.getenv("CDDTOKEN"), reason="No CDD token")
 
 
 @pytest.fixture(scope="session")
