+++ asapdiscovery-data/asapdiscovery/data/tests/test_pairs.py
@@ -3,7 +3,7 @@ from asapdiscovery.data.schema_v2.complex import Complex, PreppedComplex
 from asapdiscovery.data.schema_v2.ligand import Ligand
 from asapdiscovery.data.schema_v2.pairs import CompoundStructurePair
 from asapdiscovery.data.testing.test_resources import fetch_test_file
-from asapdiscovery.docking.docking_v2 import DockingInputPair
+from asapdiscovery.docking.docking import DockingInputPair
 
 
 @pytest.fixture(scope="session")
