+++ asapdiscovery-dataviz/asapdiscovery/dataviz/viz_v2/gif_viz.py
@@ -13,9 +13,7 @@ from asapdiscovery.data.postera.manifold_data_validation import (
 from asapdiscovery.dataviz._gif_blocks import GIFBlockData
 from asapdiscovery.dataviz.gif_viz import add_gif_progress_bar
 from asapdiscovery.dataviz.show_contacts import show_contacts
-from asapdiscovery.docking.docking_data_validation import (
-    DockingResultColsV2 as DockingResultCols,
-)
+from asapdiscovery.docking.docking_data_validation import DockingResultCols
 from asapdiscovery.simulation.simulate import SimulationResult
 from pydantic import BaseModel, Field, PositiveInt
