+++ asapdiscovery-docking/asapdiscovery/docking/tests/cli/test_docking_cli.py
@@ -18,192 +18,78 @@ def click_success(result):
 @pytest.mark.skipif(
     os.getenv("RUNNER_OS") == "macOS", reason="Docking tests slow on GHA on macOS"
 )
+@pytest.mark.skipif(os.getenv("SKIP_EXPENSIVE_TESTS"), reason="Expensive tests skipped")
 @pytest.mark.parametrize("subcommand", ["large-scale", "small-scale"])
-def test_project_support_docking_cli_fragalysis(
-    ligand_file, mpro_frag_dir, tmp_path, subcommand
-):
-    runner = CliRunner()
+class TestDockingCLICommonAPI:
+    def test_project_support_docking_cli_fragalysis(
+        self, ligand_file, mpro_frag_dir, tmp_path, subcommand
+    ):
+        runner = CliRunner()
 
-    frag_parent_dir, _ = mpro_frag_dir
+        frag_parent_dir, _ = mpro_frag_dir
 
-    args = [
-        subcommand,
-        "--target",
-        "SARS-CoV-2-Mpro",
-        "--ligands",
-        ligand_file,
-        "--fragalysis-dir",
-        frag_parent_dir,
-        "--posit-confidence-cutoff",
-        0,
-        "--output-dir",
-        tmp_path,
-    ]
-
-    if (
-        subcommand == "small-scale"
-    ):  # turn off dask cuda overrides for CI runners which lack GPUs
-        args.extend(["--no-allow-dask-cuda"])
-
-    result = runner.invoke(cli, args)
-    assert click_success(result)
-
-
-@pytest.mark.skipif(
-    os.getenv("RUNNER_OS") == "macOS", reason="Docking tests slow on GHA on macOS"
-)
-@pytest.mark.parametrize("subcommand", ["large-scale", "small-scale"])
-def test_project_support_docking_cli_structure_directory_dask(
-    ligand_file, structure_dir, tmp_path, subcommand
-):
-    runner = CliRunner()
-
-    struct_dir, _ = structure_dir
-
-    args = [
-        subcommand,
-        "--target",
-        "SARS-CoV-2-Mpro",
-        "--ligands",
-        ligand_file,
-        "--structure-dir",
-        struct_dir,
-        "--use-dask",  # add dask
-        "--posit-confidence-cutoff",
-        0,
-        "--output-dir",
-        tmp_path,
-    ]
-
-    if (
-        subcommand == "small-scale"
-    ):  # turn off dask cuda overrides for CI runners which lack GPUs
-        args.extend(["--no-allow-dask-cuda"])
-
-    result = runner.invoke(cli, args)
-    assert click_success(result)
-
-
-@pytest.mark.skipif(
-    os.getenv("RUNNER_OS") == "macOS", reason="Docking tests slow on GHA on macOS"
-)
-@pytest.mark.parametrize("subcommand", ["large-scale", "small-scale"])
-def test_project_support_docking_cli_structure_directory_du_cache_dask(
-    ligand_file, structure_dir, du_cache, tmp_path, subcommand
-):
-    runner = CliRunner()
-
-    struct_dir, _ = structure_dir
-    du_cache_dir, _ = du_cache
-
-    args = [
-        subcommand,
-        "--target",
-        "SARS-CoV-2-Mpro",
-        "--ligands",
-        ligand_file,
-        "--structure-dir",
-        struct_dir,
-        "--use-dask",
-        "--posit-confidence-cutoff",
-        0,
-        "--cache-dir",
-        du_cache_dir,
-        "--output-dir",
-        tmp_path,
-    ]
-
-    if (
-        subcommand == "small-scale"
-    ):  # turn off dask cuda overrides for CI runners which lack GPUs
-        args.extend(["--no-allow-dask-cuda"])
-
-    result = runner.invoke(cli, args)
-    assert click_success(result)
-
-
-@pytest.mark.skipif(
-    os.getenv("RUNNER_OS") == "macOS", reason="Docking tests slow on GHA on macOS"
-)
-@pytest.mark.parametrize("subcommand", ["large-scale", "small-scale"])
-def test_project_support_docking_cli_pdb_file_dask(
-    ligand_file, pdb_file, tmp_path, subcommand
-):
-    runner = CliRunner()
-
-    args = [
-        subcommand,
-        "--target",
-        "SARS-CoV-2-Mpro",
-        "--ligands",
-        ligand_file,
-        "--pdb-file",
-        pdb_file,
-        "--use-dask",
-        "--posit-confidence-cutoff",
-        0,
-        "--output-dir",
-        tmp_path,
-    ]
-
-    if (
-        subcommand == "small-scale"
-    ):  # turn off dask cuda overrides for CI runners which lack GPUs
-        args.extend(["--no-allow-dask-cuda"])
-
-    result = runner.invoke(cli, args)
-    assert click_success(result)
-
-
-@patch("asapdiscovery.simulation.simulate._SIMULATOR_TRUNCATE_STEPS", False)
-@pytest.mark.skipif(
-    os.getenv("RUNNER_OS") == "macOS", reason="Docking tests slow on GHA on macOS"
-)
-def test_small_scale_docking_md(ligand_file, pdb_file, tmp_path):
-    runner = CliRunner()
-
-    result = runner.invoke(
-        cli,
-        [
-            "small-scale",
+        args = [
+            subcommand,
             "--target",
             "SARS-CoV-2-Mpro",
             "--ligands",
             ligand_file,
-            "--pdb-file",
-            pdb_file,
-            "--use-dask",
-            "--no-allow-dask-cuda",
+            "--fragalysis-dir",
+            frag_parent_dir,
             "--posit-confidence-cutoff",
             0,
             "--output-dir",
             tmp_path,
-            "--md",
-            "--md-steps",
-            1,
-            "--md-openmm-platform",
-            "CPU",
-        ],
-    )
-    # check that the number of steps was set to 1
-    assert click_success(result)
+        ]
 
+        if (
+            subcommand == "small-scale"
+        ):  # turn off dask cuda overrides for CI runners which lack GPUs
+            args.extend(["--no-allow-dask-cuda"])
 
-@pytest.mark.skipif(
-    os.getenv("RUNNER_OS") == "macOS", reason="Docking tests slow on GHA on macOS"
-)
-def test_cross_docking_cli_structure_directory_du_cache(
-    ligand_file, structure_dir, du_cache, tmp_path
-):
-    runner = CliRunner()
+        result = runner.invoke(cli, args)
+        assert click_success(result)
 
-    struct_dir, _ = structure_dir
-    du_cache_dir, _ = du_cache
+    def test_project_support_docking_cli_structure_directory_dask(
+        self, ligand_file, structure_dir, tmp_path, subcommand
+    ):
+        runner = CliRunner()
 
-    result = runner.invoke(
-        cli,
-        [
-            "cross-docking",
+        struct_dir, _ = structure_dir
+
+        args = [
+            subcommand,
+            "--target",
+            "SARS-CoV-2-Mpro",
+            "--ligands",
+            ligand_file,
+            "--structure-dir",
+            struct_dir,
+            "--use-dask",  # add dask
+            "--posit-confidence-cutoff",
+            0,
+            "--output-dir",
+            tmp_path,
+        ]
+
+        if (
+            subcommand == "small-scale"
+        ):  # turn off dask cuda overrides for CI runners which lack GPUs
+            args.extend(["--no-allow-dask-cuda"])
+
+        result = runner.invoke(cli, args)
+        assert click_success(result)
+
+    def test_project_support_docking_cli_structure_directory_du_cache_dask(
+        self, ligand_file, structure_dir, du_cache, tmp_path, subcommand
+    ):
+        runner = CliRunner()
+
+        struct_dir, _ = structure_dir
+        du_cache_dir, _ = du_cache
+
+        args = [
+            subcommand,
             "--target",
             "SARS-CoV-2-Mpro",
             "--ligands",
@@ -211,10 +97,115 @@ def test_cross_docking_cli_structure_directory_du_cache(
             "--structure-dir",
             struct_dir,
             "--use-dask",
+            "--posit-confidence-cutoff",
+            0,
             "--cache-dir",
             du_cache_dir,
             "--output-dir",
             tmp_path,
-        ],
-    )
-    assert click_success(result)
+        ]
+
+        if (
+            subcommand == "small-scale"
+        ):  # turn off dask cuda overrides for CI runners which lack GPUs
+            args.extend(["--no-allow-dask-cuda"])
+
+        result = runner.invoke(cli, args)
+        assert click_success(result)
+
+    def test_project_support_docking_cli_pdb_file_dask(
+        self, ligand_file, pdb_file, tmp_path, subcommand
+    ):
+        runner = CliRunner()
+
+        args = [
+            subcommand,
+            "--target",
+            "SARS-CoV-2-Mpro",
+            "--ligands",
+            ligand_file,
+            "--pdb-file",
+            pdb_file,
+            "--use-dask",
+            "--posit-confidence-cutoff",
+            0,
+            "--output-dir",
+            tmp_path,
+        ]
+
+        if (
+            subcommand == "small-scale"
+        ):  # turn off dask cuda overrides for CI runners which lack GPUs
+            args.extend(["--no-allow-dask-cuda"])
+
+        result = runner.invoke(cli, args)
+        assert click_success(result)
+
+
+@pytest.mark.skipif(
+    os.getenv("RUNNER_OS") == "macOS", reason="Docking tests slow on GHA on macOS"
+)
+@pytest.mark.skipif(os.getenv("SKIP_EXPENSIVE_TESTS"), reason="Expensive tests skipped")
+class TestSmallScaleDockingMD:
+    @patch("asapdiscovery.simulation.simulate._SIMULATOR_TRUNCATE_STEPS", False)
+    def test_small_scale_docking_md(self, ligand_file, pdb_file, tmp_path):
+        runner = CliRunner()
+
+        result = runner.invoke(
+            cli,
+            [
+                "small-scale",
+                "--target",
+                "SARS-CoV-2-Mpro",
+                "--ligands",
+                ligand_file,
+                "--pdb-file",
+                pdb_file,
+                "--use-dask",
+                "--no-allow-dask-cuda",
+                "--posit-confidence-cutoff",
+                0,
+                "--output-dir",
+                tmp_path,
+                "--md",
+                "--md-steps",
+                1,
+                "--md-openmm-platform",
+                "CPU",
+            ],
+        )
+        # check that the number of steps was set to 1
+        assert click_success(result)
+
+
+@pytest.mark.skipif(
+    os.getenv("RUNNER_OS") == "macOS", reason="Docking tests slow on GHA on macOS"
+)
+@pytest.mark.skipif(os.getenv("SKIP_EXPENSIVE_TESTS"), reason="Expensive tests skipped")
+class TestCrossDockingCLI:
+    def test_cross_docking_cli_structure_directory_du_cache(
+        self, ligand_file, structure_dir, du_cache, tmp_path
+    ):
+        runner = CliRunner()
+
+        struct_dir, _ = structure_dir
+        du_cache_dir, _ = du_cache
+
+        result = runner.invoke(
+            cli,
+            [
+                "cross-docking",
+                "--target",
+                "SARS-CoV-2-Mpro",
+                "--ligands",
+                ligand_file,
+                "--structure-dir",
+                struct_dir,
+                "--use-dask",
+                "--cache-dir",
+                du_cache_dir,
+                "--output-dir",
+                tmp_path,
+            ],
+        )
+        assert click_success(result)
